assign courses to student feature

+++ eXcelerate/src/main/java/com/eXcelerate/entities/CurrentLoggedInID.java
@@ -1,6 +1,6 @@
 package com.eXcelerate.entities;
 
 public class CurrentLoggedInID {
-	public static long CurrentLoggedInStudentID;
-	public static long CurrentLoggedInInstructorsID;
+	public static int CurrentLoggedInStudentID;
+	public static int CurrentLoggedInInstructorsID;
 }
